+++ usr/include/pwd.h
@@ -1,9 +1,9 @@
 struct	passwd { /* see getpwent(3) */
 	char	*pw_name;
 	char	*pw_passwd;
-	int	pw_uid;
-	int	pw_gid;
-	int	pw_quota;
+	short	pw_uid;
+	short	pw_gid;
+	short	pw_quota;
 	char	*pw_comment;
 	char	*pw_gecos;
 	char	*pw_dir;
