+++ usr/include/sgtty.h
@@ -7,7 +7,7 @@ struct sgttyb {
 	char	sg_ospeed;		/* output speed */
 	char	sg_erase;		/* erase character */
 	char	sg_kill;		/* kill character */
-	int	sg_flags;		/* mode flags */
+	short	sg_flags;		/* mode flags */
 };
 
 /*
